--- a/src/ellipse2d.jl
+++ b/src/ellipse2d.jl
@@ -5,8 +5,6 @@ Ellipse(p::CartesianIndex{2}, args...; kwargs...) = Ellipse(Point(p), args...; k
 Ellipse(circle::CirclePointRadius) = Ellipse(circle.center, circle.ρ, circle.ρ; thickness = circle.thickness, fill = circle.fill)
 
 function draw!(img::AbstractArray{T, 2}, ellipse::Ellipse, color::T) where T<:Colorant
-	ys = Int[]
-	xs = Int[]
 	break_point = 0
 	if ellipse.fill == false
 		break_point = ((ellipse.ρy - ellipse.thickness) / ellipse.ρy) ^ 2 + ((ellipse.ρx - ellipse.thickness) / ellipse.ρx) ^ 2 
@@ -15,16 +13,14 @@ function draw!(img::AbstractArray{T, 2}, ellipse::Ellipse, color::T) where T<:Co
 		for j in ellipse.center.x : ellipse.center.x + ellipse.ρx
 			val = ((i - ellipse.center.y) / ellipse.ρy) ^ 2 + ((j - ellipse.center.x) / ellipse.ρx) ^ 2
 			if val < 1 && val >= break_point			
-				push!(ys, i)
-				push!(xs, j)
+        yi = Int(i)
+        xi = Int(j)
+        drawifinbounds!(img, yi, xi, color)
+        drawifinbounds!(img, 2 * ellipse.center.y - yi, xi, color)
+        drawifinbounds!(img, yi, 2 * ellipse.center.x - xi, color)
+        drawifinbounds!(img, 2 * ellipse.center.y - yi, 2 * ellipse.center.x - xi, color)
 			end
 		end
-	end
-	for (yi, xi) in zip(ys, xs)
-		drawifinbounds!(img, yi, xi, color)
-		drawifinbounds!(img, 2 * ellipse.center.y - yi, xi, color)
-		drawifinbounds!(img, yi, 2 * ellipse.center.x - xi, color)
-		drawifinbounds!(img, 2 * ellipse.center.y - yi, 2 * ellipse.center.x - xi, color)
 	end
 	img
 end