This PR drammatically improves on performance for ellipse drawing:

```julia
using Revise
using BenchmarkTools
using ImageDraw 
import ColorTypes: RGB, N0f8

img = rand(RGB{N0f8}, 1000, 1000)
c = CirclePointRadius(Point(500, 500), 400)
f() = draw!(img, c)
@benchmark f()
```
before this PR:
```julia
BenchmarkTools.Trial: 1875 samples with 1 evaluation.
 Range (min … max):  2.258 ms …   5.262 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     2.499 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.600 ms ± 292.306 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

       ▅██▆▂▁                                                  
  ▂▃▄▇▇███████▆▇▅▆▇██▆▅▄▃▃▃▃▃▃▃▂▂▁▂▂▂▂▂▃▃▃▃▃▃▃▂▂▂▂▂▃▃▂▂▂▂▁▁▂▂ ▃
  2.26 ms         Histogram: frequency by time        3.63 ms <

 Memory estimate: 3.66 MiB, allocs estimate: 22.
```
After this PR:
```julia
BenchmarkTools.Trial: 8135 samples with 1 evaluation.
 Range (min … max):  594.081 μs …  1.619 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     609.660 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   612.257 μs ± 26.233 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

               █        ▁                                       
  ▂▂▂▂▁▁▂▂▃▃▂▃▃█▄▄▄▇▆▅▅▄█▄▄▄▄▄▃▃▃▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ ▃
  594 μs          Histogram: frequency by time          643 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
````