@yakir12 2023-12-11
Does anyone have time to review this? The code change is pretty small and the improvement is substantial, to put it in relative terms :stuck_out_tongue: