@yakir12 2023-12-11
This file contained tab-spaces so I replaced them with (2) spaces, hope that doesn't go against the style guide of ImageDraw...